+++ src/spoke/interfaces/IBalanceSheet.sol
@@ -21,7 +21,7 @@ struct ShareQueueAmount {
     // Number of queued asset IDs for this share class
     uint32 queuedAssetCounter;
     // Nonce for share + asset messages to the hub
-    uint88 nonce;
+    uint64 nonce;
 }
 
 struct AssetQueueAmount {
@@ -67,7 +67,7 @@ interface IBalanceSheet {
     function queuedShares(PoolId poolId, ShareClassId scId)
         external
         view
-        returns (uint128 delta, bool isPositive, uint32 queuedAssetCounter, uint88 nonce);
+        returns (uint128 delta, bool isPositive, uint32 queuedAssetCounter, uint64 nonce);
     function queuedAssets(PoolId poolId, ShareClassId scId, AssetId assetId)
         external
         view
